backend/requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.5
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.3.0
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
future==0.18.2
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.2.0
jenkinsapi==0.3.11
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
msgpack==1.0.0
oauthlib==3.1.0
paramiko==2.7.1
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-telegram-bot==12.6.1
python3-openid==3.2.0
pytz==2019.3
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
websockets==8.1
zope.interface==5.1.0